fix: structure handlers into their own separate files
